fix: fixes issues with experiments not reporting desired data on UI
This also addresses issues that were inherent with the previous workflow
design. Phoenix actually doesn't run the evaluations in the same context
as the actual workflow it runs so some data we may expect to be there
isn't guaranteed to be there. To handle this we actually need to
override the tool itself and ensure we are in an environment we know
about. This isn't the ideal solution but is the one that works for now.