fix: allow semantic / raw expressions
This if a fairly serious clean up effort to remove some of the confusing
attribute resolution that was previously here. We now have explicit
SemanticExpressions or RawExpressions and expressions are now parsed
during initialization. Once parsed they're processed and any column or
semantic references are removed. This addresses issues with attempting
to reference semantic names that don't match the column names.