fix(oso_semantic): fixes attribute resolution in measures and dimensions (#4300)
* add: semantic `query` test

* fix: allow semantic / raw expressions

This if a fairly serious clean up effort to remove some of the confusing
attribute resolution that was previously here. We now have explicit
SemanticExpressions or RawExpressions and expressions are now parsed
during initialization. Once parsed they're processed and any column or
semantic references are removed. This addresses issues with attempting
to reference semantic names that don't match the column names.

* chore: remove sqlmesh as a dep from oso_semantic

* fix(oso_semantic): reenable SemanticQuery object

* fix: additional clean up

* fix tests

* Update warehouse/oso_semantic/oso_semantic/definition.py

Co-authored-by: Ícaro Guerra <icaro@plasmic.app>

* fix

---------

Co-authored-by: Reuven V. Gonzales <reuven.v.gonzales@gmail.com>
Co-authored-by: Reuven Gonzales <reuven@karibalabs.co>
Co-authored-by: Ícaro Guerra <icaro@plasmic.app>